requirements.txt
python-telegram-bot[job-queue]==21.9
google-generativeai==0.8.5
httpx==0.27.2
google-api-python-client==2.155.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
duckduckgo-search==7.3.2
apscheduler==3.10.4
python-dotenv==1.0.1
aiosqlite==0.20.0
python-dateutil==2.9.0
